+++ OrderProcessingDomain/Repositories/IRepository.cs
@@ -16,6 +16,6 @@ public interface IRepository<T> where T : class
   void Save(object toSave);
   void Update(object toUpdate);
   void RemoveWhere(Func<T, bool> exp);
-  int Count();
+  Int64 Count();
   T Get<T>(object key);
 }
